Add files via upload
Made the GUI.  Did not implement any of the functionality, So buttons and text fields don't do anything yet, but it looks super.  Also realized that we did not need all the JLabels to be instance variables, so i got rid of some.  I figure only only the output labels that update as the sim runs need to be instance variables.