Add files via upload
Need two of these to be fed by VoterProducer. Both of them will feed the booth. I am confused as to how the destination stuff works. I have it so that you must pass this class booth and at the end of event it removes the person in the front of the Q and adds them to booth. I think this should work but I am not sure.