Added different voter types.
Pretty sure this is working.  My algorithm for determining the voters is real simple, just randomly generate a number from 0-9 and assign them to the different voter types in voter producer. feel free to change it if you have a better idea.  but i think it works. give it a look. I have no idea how its supposed to change the numbers.  It speeds up the average total voter time, but i think thats good because 20% of the voters are faster. but check it out.  seems to have good SKEWHGE